requirements.txt
accelerate
codetiming
datasets
flash-attn>=2.4.3
liger-kernel
mathruler
numpy
omegaconf
pandas
peft
pillow
pyarrow>=15.0.0
pylatexenc
qwen-vl-utils
ray[default]
tensordict
torchdata
transformers==4.51.3
vllm==0.8.2
wandb
tensorboard
gradio==4.44.1
pydantic==2.10.6
tabulate